requirements.txt
hyperliquid-python-sdk>=0.1.0
eth-account>=0.8.0
web3>=6.0.0
requests>=2.28.0
websocket-client>=1.5.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
ccxt>=4.0.0

